Creación de Subcategorías en Server y BDD

+++ backend/src/controllers/Products/removeProduct.js
@@ -1,6 +1,14 @@
 const { Product } = require("../../db");
+const findProductbyId = require("./findProductbyId");
 
 const removeProduct = async (id) => {
+  const product = await findProductbyId(id);
+  if (!product) {
+    return {
+      message: `No existe un Producto con id '${id}' para eliminar`,
+    };
+  }
+
   try {
     productDestroyed = await Product.destroy({
       where: { id: id },
